Put CapitalizeDecorator and TrimmerDecorator into two separate files, as required by code reviewer
