+++ capitalizeDecorator.rb
@@ -1,0 +1,12 @@
+require_relative 'decorator'
+
+class CapitalizeDecorator < Decorator
+  def initialize(nameable)
+    super(nameable)
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
