+++ decorator.rb
@@ -11,24 +11,3 @@ class Decorator < Nameable
   end
 end
 
-class CapitalizeDecorator < Decorator
-  def initialize(nameable)
-    super(nameable)
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator
-  def initialize(nameable)
-    # super(nameable)
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name.length > 10 ? @nameable.correct_name[0...10] : @nameable.correct_name
-  end
-end
